GH-48612: [Ruby] Add support for reading streaming format
